+++ lib/provider/list_provider.dart
@@ -1,29 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:to_do_app/UI/home_screen/tabs/list_tabs/todo_wodget.dart';
-import 'package:to_do_app/model/todo_dm.dart';
 
-class ListTab extends StatefulWidget {
-  // static const routeNmae = 'ListTab';
+import '../model/todo_dm.dart';
 
-  @override
-  State<ListTab> createState() => _ListTabState();
-}
-
-class _ListTabState extends State<ListTab> {
+class ListProvider extends ChangeNotifier {
   List<ToDoDM> todosList = [];
-
-  @override
-  Widget build(BuildContext context) {
-    if (todosList.isEmpty) refreshTodosFromFireStore();
-    return ListView.builder(
-      itemCount: todosList.length,
-      itemBuilder: (context, index) {
-        return ToDo(todo: todosList[index]);
-      },
-    );
-  }
-
   refreshTodosFromFireStore() {
     var todosCollection = FirebaseFirestore.instance.collection("todos");
     // .then : wait datafrom firestore
@@ -37,7 +18,7 @@ class _ListTabState extends State<ListTab> {
             isDone: json["isDone"],
             date: DateTime.fromMicrosecondsSinceEpoch(json["datetime"]));
       }).toList();
-      setState(() {});
+      notifyListeners();
     });
   }
 }
