+++ lib/screens/home_screen/tabs/list_tabs/todo_wodget.dart
@@ -1,7 +1,10 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_swipe_action_cell/core/cell.dart';
-import 'package:to_do_app/UI/theme/app_colors.dart';
+import 'package:provider/provider.dart';
 import 'package:to_do_app/model/todo_dm.dart';
+import 'package:to_do_app/provider/list_provider.dart';
+import 'package:to_do_app/screens/theme/app_colors.dart';
 
 class ToDo extends StatelessWidget {
   ToDoDM todo;
@@ -10,6 +13,7 @@ class ToDo extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    ListProvider provider = Provider.of(context);
     return Container(
       margin: EdgeInsets.symmetric(horizontal: 30, vertical: 22),
       // padding: EdgeInsets.symmetric(horizontal: 18, vertical: 25),
@@ -19,7 +23,17 @@ class ToDo extends StatelessWidget {
         key: ObjectKey(todo),
         leadingActions: [
           SwipeAction(
-              onTap: (handler) {},
+              onTap: (handler) {
+                var todosCollection =
+                    FirebaseFirestore.instance.collection("todos");
+                var doc = todosCollection.doc(todo.id);
+                doc.delete().timeout(
+                  Duration(milliseconds: 500),
+                  onTimeout: () {
+                    provider.refreshTodosFromFireStore();
+                  },
+                );
+              },
               backgroundRadius: 25,
               content: Container(
                 child: Column(
