+++ lib/screens/home_screen/tabs/settings/setting_tab.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:to_do_app/screens/theme/app_colors.dart';
 
 class SettingTab extends StatelessWidget {
   static const routeNmae = 'SettingTab';
@@ -6,7 +7,7 @@ class SettingTab extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      color: Colors.amber,
+      color: AppColors.accentColor,
     );
   }
 }
